Extract instance properties schema making logic in dedicated file
